self_improving_rag/pyproject.toml
[project]
name = "self-improving-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain==1.2.15",
    "langgraph==1.1.0",
    "langchain_community==0.4.1",
    "langchain_openai==1.2.1",
    "langchain_core==1.3.2",
    "ollama==0.4.5",
    "pandas==2.2.3",
    "duckdb==1.1.0",
    "faiss-cpu==1.10.0",
    "sentence-transformers==3.4.1",
    "biopython==1.85",
    "pypdf==5.3.0",
    "pydantic==2.10.6",
    "lxml==5.3.1",
    "html2text==2024.2.26",
    "beautifulsoup4==4.13.3",
    "matplotlib==3.10.1",
]
